Refactor path element to store under package path
